extract converter logic.
